requirements.txt
beautifulsoup4==4.6.0
certifi==2017.4.17
chardet==3.0.4
click==6.7
colorlover==0.2.1
dash-core-components==0.5.0
dash-html-components==0.6.0
dash-renderer==0.7.1
dash.ly==0.17.0
decorator==4.0.11
enum34==1.1.6
finsymbols==1.2.0
Flask==0.12.2
Flask-Compress==1.4.0
Flask-SeaSurf==0.2.2
functools32==3.2.3.post2
gunicorn==19.7.1
idna==2.5
ipython-genutils==0.2.0
itsdangerous==0.24
Jinja2==2.9.6
jsonschema==2.6.0
jupyter-core==4.3.0
MarkupSafe==1.0
nbformat==4.3.0
numpy==1.13.0
pandas==0.20.2
pandas-datareader==0.4.0
plotly==2.0.9
python-dateutil==2.6.0
pytz==2017.2
requests==2.18.1
requests-file==1.4.2
requests-ftp==0.3.1
six==1.10.0
traitlets==4.3.2
urllib3==1.21.1
Werkzeug==0.12.2
wheel==0.26.0
